requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
blis==1.1.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.0.3
cloudpathlib==0.20.0
colorama==0.4.4
comm==0.2.2
confection==0.1.5
constantly==15.1.0
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.10
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
distro==1.7.0
efficientnet-pytorch==0.7.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu==1.7.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hilbertcurve==2.0.5
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
imageio==2.35.1
incremental==21.3.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
keras-models==0.0.7
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
mahotas==1.4.18
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
murmurhash==1.0.11
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openslide-python==1.3.1
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.3.6
portalocker==2.10.1
preshed==3.0.9
pretrainedmodels==0.7.4
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-triton==3.1.0+5fe38ffd73
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.1
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
segmentation-models-pytorch==0.3.4
service-identity==18.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
ssh-import-id==5.11
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==8.3.3
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.7
torch==2.6.0.dev20240920+cu121
torchaudio==2.5.0.dev20240920+cu121
torchvision==0.20.0.dev20240920+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
Twisted==22.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
yacs==0.1.8
yarl==1.11.1
zope.interface==5.4.0
